chore(suite-common) Device types moved to common package
