+++ packages/suite-build/configs/web.webpack.config.ts
@@ -4,7 +4,7 @@ import CopyPlugin from 'copy-webpack-plugin';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
 
-import routes from '../../suite/src/config/suite/routes';
+import routes from '../../../suite-common/suite-config/src/routes';
 import { FLAGS } from '../../suite/src/config/suite/features';
 
 import { assetPrefix, isDev } from '../utils/env';
