+++ packages/suite/package.json
@@ -38,7 +38,10 @@
         "@trezor/validation": "*",
         "@suite-common/intl-types": "*",
         "@suite-common/metadata-types": "*",
-        "@suite-common/suite-features-config": "*",
+        "@suite-common/suite-config": "*",
+        "@suite-common/suite-constants": "*",
+        "@suite-common/suite-support": "*",
+        "@suite-common/suite-types": "*",
         "@suite-common/suite-utils": "*",
         "@suite-common/wallet-networks-config": "*",
         "@suite-common/wallet-types": "*",
