+++ packages/suite/src/actions/suite/__tests__/routerActions.test.ts
@@ -74,7 +74,8 @@ describe('Suite Actions', () => {
             const state = getInitialState(f.state as InitialState);
             const store = initStore(state);
             // eslint-disable-next-line global-require
-            require('@suite/support/history').default.location.pathname = f.pathname || '/';
+            require('@suite-common/suite-support').history.default.location.pathname =
+                f.pathname || '/';
             store.dispatch(routerActions.initialRedirection());
             expect(store.getState().router.app).toEqual(f.app);
         });
@@ -85,7 +86,7 @@ describe('Suite Actions', () => {
             const state = getInitialState(f.state as InitialState);
             const store = initStore(state);
             // eslint-disable-next-line global-require
-            require('@suite/support/history').default.location.hash = `#${f.hash}`;
+            require('@suite-common/suite-support').history.default.location.hash = `#${f.hash}`;
             store.dispatch(routerActions.goto(f.url as any, { preserveParams: f.preserveHash }));
             if (f.result) {
                 expect(store.getActions()[0].payload.url).toEqual(f.result);
@@ -112,7 +113,7 @@ describe('Suite Actions', () => {
         const state = getInitialState({ router: { pathname: '/firmware' } });
         const store = initStore(state);
         // eslint-disable-next-line global-require
-        require('@suite/support/history').default.location.pathname = '/accounts/send';
+        require('@suite-common/suite-support').history.default.location.pathname = '/accounts/send';
         store.dispatch(routerActions.closeModalApp());
         expect(store.getActions().length).toEqual(2); // unlock + location change
         expect(store.getState().router.app).toEqual('wallet');
