+++ packages/suite/src/actions/suite/routerActions.ts
@@ -3,7 +3,7 @@
  */
 import * as suiteActions from '@suite-actions/suiteActions';
 import { SUITE, ROUTER } from '@suite-actions/constants';
-import { RouterAppWithParams, SettingsBackRoute } from '@suite-constants/routes';
+import { RouterAppWithParams, SettingsBackRoute, RouteParams } from '@suite-common/suite-types';
 import {
     getAppWithParams,
     getPrefixedURL,
@@ -11,10 +11,9 @@ import {
     getBackgroundRoute,
     findRoute,
     findRouteByName,
-    RouteParams,
-} from '@suite-utils/router';
+} from '@suite-common/suite-utils';
 import { Dispatch, GetState, Route } from '@suite-types';
-import history from '@suite/support/history';
+import { history } from '@suite-common/suite-support';
 
 import type { AnchorType } from '@suite-constants/anchors';
