+++ packages/suite/src/types/wallet/index.ts
@@ -42,8 +42,11 @@ export type DiscoveryStatus =
               | 'discovery-failed'
               | 'device-unavailable';
       };
-export type { WalletParams } from '@suite-utils/router';
-export type { WalletAccountTransaction, RbfTransactionParams } from '@suite-common/wallet-types';
+export type {
+    WalletAccountTransaction,
+    RbfTransactionParams,
+    WalletParams,
+} from '@suite-common/wallet-types';
 export type { ReceiveInfo } from '@wallet-reducers/receiveReducer';
 
 export type WalletAction =
