+++ packages/suite/src/utils/suite/oauth.ts
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
 import { desktopApi } from '@trezor/suite-desktop-api';
-import { getPrefixedURL } from '@suite-utils/router';
+import { getPrefixedURL } from '@suite-common/suite-utils';
 import { METADATA } from '@suite-actions/constants';
 import { Deferred, createDeferred } from '@trezor/utils';
 import { urlHashParams, urlSearchParams } from '@suite-utils/metadata';
