+++ suite-common/suite-constants/src/routes.ts
@@ -1,0 +1,4 @@
+import { Route } from '@suite-common/suite-types';
+import { routes } from '@suite-common/suite-config';
+
+export default [...routes] as Route[];
