+++ suite-common/suite-constants/tsconfig.json
@@ -1,5 +1,8 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": []
+    "references": [
+        { "path": "../suite-config" },
+        { "path": "../suite-types" }
+    ]
 }
