+++ suite-common/suite-support/src/index.ts
@@ -1,0 +1,1 @@
+export { default as history } from './history';
