+++ suite-common/suite-types/package.json
@@ -11,6 +11,10 @@
     },
     "dependencies": {
         "@suite-common/metadata-types": "*",
-        "@trezor/connect": "9.0.0-beta.4"
+        "@suite-common/suite-config": "*",
+        "@suite-common/wallet-networks-config": "*",
+        "@suite-common/wallet-types": "*",
+        "@trezor/connect": "9.0.0-beta.4",
+        "@trezor/type-utils": "*"
     }
 }
