+++ suite-common/suite-types/src/index.ts
@@ -1,1 +1,2 @@
 export * from './device';
+export * from './routes';
