+++ suite-common/suite-utils/src/index.ts
@@ -1,1 +1,2 @@
 export * from './features';
+export * from './router';
