+++ suite-common/wallet-types/src/routes.ts
@@ -1,0 +1,4 @@
+import { validateModalAppParams, validateWalletParams } from '@suite-common/suite-utils';
+
+export type WalletParams = NonNullable<ReturnType<typeof validateWalletParams>>;
+export type ModalAppParams = NonNullable<ReturnType<typeof validateModalAppParams>>;
